- Add lib ViewThreadDecorator - Remove includes presentation-common and presentation-compiler from gradle.build - Move Presenter, InteractorInvoker, InteractorOutput and Qualifiers to presentation package -
