+++ app/src/main/java/me/panavtec/cleancontacts/domain/outputs/BackThreadSpec.java
@@ -1,6 +1,6 @@
 package me.panavtec.cleancontacts.domain.outputs;
 
-import me.panavtec.presentation.common.ThreadSpec;
+import me.panavtec.threaddecoratedview.views.ThreadSpec;
 
 public class BackThreadSpec implements ThreadSpec {
