+++ build.gradle
@@ -6,7 +6,7 @@ buildscript {
     mavenCentral()
   }
   dependencies {
-    classpath 'com.android.tools.build:gradle:1.3.0'
+    classpath 'com.android.tools.build:gradle:1.3.1'
     classpath 'com.neenbedankt.gradle.plugins:android-apt:1.8'
   }
 }
