+++ presentation/build.gradle
@@ -8,8 +8,8 @@ targetCompatibility = JavaVersion.VERSION_1_7
 
 dependencies {
   compile project(':domain')
-  compile project(':presentation-common')
-  apt project(":presentation-compiler")
+  compile 'me.panavtec:threaddecoratedview-common:1.5.1'
+  compile 'me.panavtec:threaddecoratedview-compiler:1.5.1'
 
   testCompile rootProject.ext.testDependencies.junit
   testCompile rootProject.ext.testDependencies.mockito
