+++ presentation/src/main/java/me/panavtec/cleancontacts/presentation/invoker/InteractorInvoker.java
@@ -1,6 +1,7 @@
 package me.panavtec.cleancontacts.presentation.invoker;
 
 import java.util.concurrent.Future;
+
 import me.panavtec.cleancontacts.domain.interactors.Interactor;
 import me.panavtec.presentation.common.outputs.InteractorOutput;
 
@@ -11,8 +12,8 @@ public interface InteractorInvoker {
   <T, E extends Exception> Future<T> execute(Interactor<T, E> interactor, int priority);
 
   <T, E extends Exception> Future<T> execute(Interactor<T, E> interactor,
-      InteractorOutput<T, E> output);
+                                             InteractorOutput<T, E> output);
 
   <T, E extends Exception> Future<T> execute(Interactor<T, E> interactor,
-      InteractorOutput<T, E> output, int priority);
-}
+                                             InteractorOutput<T, E> output, int priority);
+  }
