+++ presentation/src/main/java/me/panavtec/cleancontacts/presentation/modules/detail/DetailView.java
@@ -1,13 +1,14 @@
 package me.panavtec.cleancontacts.presentation.modules.detail;
 
 import me.panavtec.cleancontacts.presentation.model.PresentationContact;
-import me.panavtec.presentation.common.views.qualifiers.NoDecorate;
-import me.panavtec.presentation.common.views.qualifiers.ThreadDecoratedView;
+import me.panavtec.threaddecoratedview.views.qualifiers.NotDecorated;
+import me.panavtec.threaddecoratedview.views.qualifiers.ThreadDecoratedView;
+
 
 @ThreadDecoratedView
 public interface DetailView {
 
-  @NoDecorate void initUi();
+  @NotDecorated void initUi();
 
   void showContactData(PresentationContact contact);
