+++ presentation/src/main/java/me/panavtec/cleancontacts/presentation/modules/main/MainPresenter.java
@@ -1,6 +1,7 @@
 package me.panavtec.cleancontacts.presentation.modules.main;
 
 import java.util.List;
+
 import me.panavtec.cleancontacts.domain.entities.Contact;
 import me.panavtec.cleancontacts.domain.interactors.contacts.GetContactsInteractor;
 import me.panavtec.cleancontacts.domain.interactors.contacts.exceptions.RetrieveContactsException;
@@ -8,12 +9,12 @@ import me.panavtec.cleancontacts.presentation.invoker.InteractorInvoker;
 import me.panavtec.cleancontacts.presentation.model.PresentationContact;
 import me.panavtec.cleancontacts.presentation.model.mapper.base.ListMapper;
 import me.panavtec.presentation.Presenter;
-import me.panavtec.presentation.common.ThreadSpec;
 import me.panavtec.presentation.common.outputs.InteractorOutput;
 import me.panavtec.presentation.common.outputs.InteractorOutputInjector;
 import me.panavtec.presentation.common.outputs.qualifiers.OnError;
 import me.panavtec.presentation.common.outputs.qualifiers.OnResult;
 import me.panavtec.presentation.common.outputs.qualifiers.Output;
+import me.panavtec.threaddecoratedview.views.ThreadSpec;
 
 public class MainPresenter extends Presenter<MainView> {
   private final InteractorInvoker interactorInvoker;
