+++ presentation/src/main/java/me/panavtec/presentation/common/ThreadSpec.java
