+++ presentation/src/main/java/me/panavtec/presentation/common/outputs/InteractorOutput.java
