+++ presentation/src/main/java/me/panavtec/presentation/common/outputs/qualifiers/OnResult.java
