+++ presentation/src/main/java/me/panavtec/presentation/common/outputs/qualifiers/Output.java
